Break `test_child_procs` down
tests/test_child_procs/
    Separated `tests.py` into these files:
    - `_test_child_procs_utils.py`:
      Util constants, functions and classes used by the tests and the
      fixtures
    - `multiproc_examples/*.py`:
      Example scripts that are profiled
    - `conftest.py`:
      Fixtures and fixture functions
    - `test_child_procs.py`:
      Actual tests

TODO: Fixes; new tests for non-`Pool`-based parallelism
